+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheetAppearance.swift
@@ -295,6 +295,9 @@ public extension PaymentSheet.Appearance {
                 case floatingButton
                 /// A flat style with a checkmark
                 case flatWithCheckmark
+                /// A flat style with a chevron
+                /// Note that EmbeddedPaymentElement.Configuration.RowSelectionBehavior must be set to `immediateAction` to use this style.
+                case flatWithChevron
             }
 
             /// The display style of the row
@@ -320,7 +323,7 @@ public extension PaymentSheet.Appearance {
                 public var separatorColor: UIColor?
 
                 /// The insets of the separator line between rows
-                /// - Note: If `nil`, defaults to `UIEdgeInsets(top: 0, left: 30, bottom: 0, right: 0)` for style of `flatWithRadio` and to `UIEdgeInsets.zero` for style of `flatWithCheckmark`.
+                /// - Note: If `nil`, defaults to `UIEdgeInsets(top: 0, left: 30, bottom: 0, right: 0)` for style of `flatWithRadio` and to `UIEdgeInsets.zero` for styles of `flatWithCheckmark` and `flatWithChevron`.
                 public var separatorInsets: UIEdgeInsets?
 
                 /// Determines if the top separator is visible at the top of the Embedded Mobile Payment Element
@@ -334,6 +337,9 @@ public extension PaymentSheet.Appearance {
 
                 /// Appearance settings for the checkmark
                 public var checkmark: Checkmark = Checkmark()
+
+                /// Appearance settings for the chevron
+                public var chevron: Chevron = Chevron()
 
                 /// Describes the appearance of the radio button
                 public struct Radio: Equatable {
@@ -351,6 +357,13 @@ public extension PaymentSheet.Appearance {
                     /// The color of the checkmark button when selected
                     /// - Note: If `nil`, defaults to `appearance.color.primaryColor`
                     public var color: UIColor?
+                }
+
+                /// Describes the appearance of the chevron
+                /// Note that EmbeddedPaymentElement.Configuration.RowSelectionBehavior must be set to `immediateAction` to use this style.
+                public struct Chevron: Equatable {
+                    /// The color of the chevron icon
+                    public var color: UIColor = .systemGray
                 }
             }